Split the file into two parts: abstracts and concretes
